+++ src/USA/MoMoe/MentionMe/ReloadCommand.java
@@ -30,5 +30,4 @@ public class ReloadCommand implements CommandExecutor {
             return true;
         }
     }
-    
 }